Split `sqlalchemy_utils` into two modules
